% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thyroidBRS-package.R
\docType{package}
\name{thyroidBRS-package}
\alias{thyroidBRS}
\alias{thyroidBRS-package}
\title{thyroidBRS: BRAF-RAS Score Classifier for Papillary Thyroid Carcinoma}
\description{
Computes the BRAF-RAS Score (BRS) of Agrawal et al. (2014) for papillary
thyroid carcinoma samples from gene expression alone, using reference
centroids fit from any cohort with known BRAF-mutant/RAS-mutant labels.
See \code{\link[=brs_fit]{brs_fit()}} to get started.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/camilasauria/thyroidBRS}
  \item Report bugs at \url{https://github.com/camilasauria/thyroidBRS/issues}
}

}
\author{
\strong{Maintainer}: Hugo Tovar \email{hatovar@inmegen.gob.mx} (\href{https://orcid.org/0000-0002-8360-6133}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Camila Barrios-Cervantes \email{paulacamilabarrioscervantes@gmail.com} (\href{https://orcid.org/0009-0006-9971-790X}{ORCID}) [copyright holder]
  \item Daniel Pérez-Calixto \email{daniel_perez@ciencias.unam.mx} (\href{https://orcid.org/0009-0004-3988-1260}{ORCID}) [copyright holder]
}

}
\keyword{internal}
